- restructured the src directories (kasa & khaya) to have khaya (now the util dir) as a subdirectory of kasa - refactoring and changes with imports - performed tests (all 22 passed with 6 httpx deprecation warnings) - identified tests fails a throws "Too Many Requests" error after running tests too closely
