+++ src/kasa/__init__.py
@@ -1,0 +1,3 @@
+from .khaya_client import KhayaClient
+
+__all__ = ["KhayaClient"]
