+++ src/kasa/constants.py
