+++ src/kasa/exceptions.py
