+++ src/kasa/khaya/text_chunker.py
