+++ src/kasa/services/tts.py
@@ -2,9 +2,9 @@ import json
 
 from requests.models import Response
 
-from src.khaya.services.base_api import BaseApi
-from src.khaya.exceptions import TTSGenerationError
-from src.khaya.utils import check_authentication
+from kasa.services.base_api import BaseApi
+from kasa.exceptions import TTSGenerationError
+from kasa.utils import check_authentication
 
 
 class TtsService:
