--- src/khaya/__init__.py
@@ -1,3 +1,0 @@
-from .khaya_client import KhayaClient
-
-__all__ = ["KhayaClient"]
